This commit has changes to `dbt-create` and `dbt-workarea-env`.
It allows developers to skip the creation of python venv (either via clone or fresh install) inside a workarea when using `dbt-create`.

Unless `-i/--install-pyvenv` or `-c/--clone-pyvenv` is used with `dbt-create`, the workarea will not contain a pre-installed python venv.

`dbt-workarea-env` will check the existence of `$DBT_WORKAREA_ROOT/$DBT_VENV`, if not found, it will activate the python venv in the release directory on cvmfs.