Adding support for creating local spack instance inside a workarea.
Use option `-s/--spack` when calling `dbt-create`.

This optionw will create `.spack` under the workarea, and add `LOCAL_SPACK_DIR` env to the `dbt-workarea-constants.sh`.

The local `.spack` is a clone of the spack instance from the release directory on cvmfs with the following modifications:
1. the packages installed in the release are skipped;
2. a `spack-repo` subdir is created, and added to the `repos.yaml` in the local `.spack`. In case one need to modify a package's recipe file, they should copy over the existing recipe file into `.spack/spack-repo/packages/<pkg_name>`, and modify it there;
3. the spack instance from the release is added as an upstream spack instance, together with all existing upstream instances used by the release.

When calling `dbt-workarea-env`, it will check if `LOCAL_SPACK_DIR` is set, if so, load the spack setup script from there.